Revert "listing fixes"
