+++ src/components/DelegationDashboard/ProfileBalance.tsx
@@ -19,7 +19,6 @@ import { APPNAME } from '~src/global/appName';
 import getEncodedAddress from '~src/util/getEncodedAddress';
 import { formatBalance } from '@polkadot/util';
 import Image from 'next/image';
-import { formatedBalance } from '~src/util/formatedBalance';
 
 interface Props{
   className?: string;
@@ -29,6 +28,16 @@ interface Props{
 const DelegationWalletConnectModal = dynamic(() => import('./DelegationWalletConnectModal'), {
 	ssr: false
 });
+
+export const formatedBalance = (balance: string, unit: string) => {
+	const formated = formatBalance(balance, { forceUnit: unit, withUnit: false }).split('.');
+	if(Number(formated?.[0][0]) > 0){
+		return formated?.[1] ? `${formated[0]}.${formated[1].slice(0,1)}`: `${formated[0]}`;
+	}else{
+		return formated.join('.');
+	}
+
+};
 
 const ProfileBalances = ({ className, address }: Props ) => {
