+++ src/components/Listing/Tracks/TrackListingAllTabContent.tsx
@@ -54,10 +54,6 @@ const TrackListingAllTabContent: FC<ITrackListingAllTabContentProps> = (props) =
 										tags={post?.tags}
 										requestedAmount={post?.requestedAmount}
 										spam_users_count={post?.spam_users_count}
-										tally={post?.tally}
-										timeline={post?.timeline || []}
-										statusHistory={post?.status_history || []}
-										index={index}
 									/>
 								</Link>}
 							</div>
