+++ src/components/Listing/Tracks/TrackListingStatusTabContent.tsx
@@ -60,10 +60,6 @@ const TrackListingStatusTabContent: FC<ITrackListingStatusTabContentProps> = (pr
 										tags={post?.tags}
 										requestedAmount={post?.requestedAmount}
 										spam_users_count={post.spam_users_count}
-										tally={post?.tally}
-										timeline={post?.timeline || []}
-										statusHistory={post?.status_history || []}
-										index={index}
 									/>
 								</Link>}
 							</div>
